Basic restucture done
