+++ Example/README.md
@@ -1,0 +1,6 @@
+# Example
+These examples have run successfully inside of the Spyder notebook.
+
+If you try to run from the command line, you may have issues with finding and loading modules
+
+
