+++ Example/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+# this file intentionally blank
+
