+++ __init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+# This file intentionally left blank
